backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.32.3
gunicorn==22.0.0
python-dotenv==1.0.1
